Rename project from tiny_tokenizer to konoha
