+++ konoha/word_tokenizers/sentencepiece_tokenizer.py
@@ -1,5 +1,5 @@
-from tiny_tokenizer.tiny_tokenizer_token import Token
-from tiny_tokenizer.word_tokenizers.tokenizer import BaseTokenizer
+from konoha.konoha_token import Token
+from konoha.word_tokenizers.tokenizer import BaseTokenizer
 
 
 class SentencepieceTokenizer(BaseTokenizer):
