+++ pyproject.toml
@@ -1,5 +1,5 @@
 [tool.poetry]
-name = "tiny_tokenizer"
+name = "konoha"
 version = "3.1.0"
 description = "A tiny sentence/word tokenizer for Japanese text written in Python"
 authors = ["himkt <himkt@cookpad.com>"]
