+++ tests/word_tokenizer/test_segmentation.py
@@ -3,8 +3,8 @@ import unittest
 
 import pytest
 
-from tiny_tokenizer.tiny_tokenizer_token import Token
-from tiny_tokenizer.word_tokenizer import WordTokenizer
+from konoha.konoha_token import Token
+from konoha.word_tokenizer import WordTokenizer
 
 SENTENCE1 = "吾輩は猫である"
 SENTENCE2 = "医薬品安全管理責任者"
