+++ tests/word_tokenizer/test_segmentation_with_remote.py
@@ -1,7 +1,7 @@
 import pytest
 
-from tiny_tokenizer.tiny_tokenizer_token import Token
-from tiny_tokenizer.word_tokenizer import WordTokenizer
+from konoha.konoha_token import Token
+from konoha.word_tokenizer import WordTokenizer
 
 SENTENCE1 = "吾輩は猫である"
